Moving images into one location staring with admin
